+++ tests/unit-mcdc/test_wc_xmss_impl_whitebox.c
@@ -207,10 +207,860 @@ static void wb_run(void)
 
 #endif
 
+#ifdef WOLFSSL_HAVE_XMSS
+
+/********************************************
+ * Shared small-parameter / state helpers
+ ********************************************/
+
+/* Hand-build an XmssParams the same way wc_xmss.c's XMSS_PARAMS() macro
+ * would (that macro itself is not visible here - it lives in wc_xmss.c),
+ * but with a caller-chosen (deliberately tiny) height/depth so full
+ * keygen/sign/verify cycles are cheap enough to run repeatedly. bds_k must
+ * keep (sub_h - bds_k) even/sane for the BDS bookkeeping; 0 is always safe.
+ */
+static void wb_params_init(XmssParams* p, byte hash, byte n, byte pad_len,
+    byte h, byte d, byte idx_len, byte bds_k)
+{
+    byte sub_h = (byte)(h / d);
+    word8 hsk = (word8)(sub_h - bds_k);
+
+    XMEMSET(p, 0, sizeof(*p));
+    p->hash = hash;
+    p->n = n;
+    p->pad_len = pad_len;
+    p->wots_len = (word8)(n * 2 + 3);
+    p->wots_sig_len = (word16)(n * p->wots_len);
+    p->h = h;
+    p->sub_h = sub_h;
+    p->d = d;
+    p->idx_len = idx_len;
+    p->sig_len = (word32)idx_len + n +
+        (word32)d * ((word32)n * 2 + 3) * n + (word32)h * n;
+    /* sk_len: replicate XMSS_SK_LEN(n,h,d,sub_h,idx_len,bds_k)'s formula
+     * from wc_xmss.c (not visible to this TU). Callers additionally
+     * over-allocate their sk buffers well beyond this. */
+    p->sk_len = (word32)idx_len + 4U * n +
+        (word32)(2 * d - 1) * ((word32)(sub_h + 1) * n + (word32)(sub_h + 1) +
+            (word32)sub_h * n + (word32)(sub_h >> 1) * n +
+            (word32)hsk * 4U + (word32)hsk * n +
+            XMSS_RETAIN_LEN(bds_k, n) + 4U) +
+        (word32)(d - 1) * n * ((word32)n * 2 + 3);
+    p->pk_len = (word8)(n * 2);
+    p->bds_k = bds_k;
+}
+
+/* Initialize an XmssState's digest for the hash family named in params.
+ * Returns 0 on success, matching wc_xmss_digest_init()'s own contract
+ * (which is file-static in wc_xmss.c and not reachable from here). */
+static int wb_state_init(XmssState* state, const XmssParams* params)
+{
+    int ret;
+
+    XMEMSET(state, 0, sizeof(*state));
+    state->params = params;
+    state->heap = NULL;
+    state->ret = 0;
+
+#ifdef WC_XMSS_SHA256
+    if (params->hash == WC_HASH_TYPE_SHA256) {
+        ret = wc_InitSha256(&state->digest.sha256);
+    }
+    else
+#endif
+#ifdef WC_XMSS_SHA512
+    if (params->hash == WC_HASH_TYPE_SHA512) {
+        ret = wc_InitSha512(&state->digest.sha512);
+    }
+    else
+#endif
+    {
+        ret = WC_NO_ERR_TRACE(NOT_COMPILED_IN);
+    }
+
+    return ret;
+}
+
+static void wb_state_free(XmssState* state)
+{
+#ifdef WC_XMSS_SHA256
+    if (state->params->hash == WC_HASH_TYPE_SHA256) {
+        wc_Sha256Free(&state->digest.sha256);
+        return;
+    }
+#endif
+#ifdef WC_XMSS_SHA512
+    if (state->params->hash == WC_HASH_TYPE_SHA512) {
+        wc_Sha512Free(&state->digest.sha512);
+        return;
+    }
+#endif
+    (void)state;
+}
+
+/********************************************
+ * 614/615, 1033-1035, 1218-1220, 1813-1815, 1884-1886, 1952-1954,
+ * 2022-2024, 2053-2055:
+ *   "params->n == <32-bit-digest-size>" (held alongside pad_len==32 and/or
+ *   hash==SHA256, whichever the decision requires) selects the SHA-256/
+ *   32-byte fast path vs. the fully generic path. Driving the SAME helper
+ *   with n=32 (fast path taken) and n=24 (still SHA-256, still within the
+ *   192..256-bit partial-digest range, but NOT 32 bytes, so the fast path's
+ *   condition is false and the generic/partial path runs) isolates the "n"
+ *   operand while hash==SHA256 (and pad_len==32, where applicable) stay
+ *   true in both calls.
+ ********************************************/
+static void wb_hash_family_pairs(void)
+{
+    XmssParams paramsFull;
+    XmssParams paramsPartial;
+    XmssState state;
+    HashAddress addr;
+    byte sk_seed[32];
+    byte pk_seed[32];
+    byte data[64];
+    byte hashOut[64];
+    byte pkBuf[WC_XMSS_MAX_WOTS_SIG_LEN];
+    byte sigBuf[WC_XMSS_MAX_WOTS_SIG_LEN];
+    int i;
+
+    wb_params_init(&paramsFull, WC_HASH_TYPE_SHA256, 32, 32, 4, 1, 4, 0);
+    wb_params_init(&paramsPartial, WC_HASH_TYPE_SHA256, 24, 32, 4, 1, 4, 0);
+
+    if (wb_state_init(&state, &paramsFull) != 0) {
+        WB_NOTE("hash family pairs: SHA-256 state init failed; skipped");
+        return;
+    }
+
+    for (i = 0; i < 64; i++) {
+        data[i] = (byte)(i ^ 0x5a);
+    }
+    XMEMSET(sk_seed, 0x77, sizeof(sk_seed));
+    XMEMSET(pk_seed, 0x88, sizeof(pk_seed));
+    XMEMSET(hashOut, 0, sizeof(hashOut));
+
+    /* Line 614/615: wc_xmss_hash()'s "params->n == WC_SHA256_DIGEST_SIZE"
+     * operand. */
+    state.ret = 0;
+    wc_xmss_hash(&state, data, 16, hashOut);
+    if (state.ret != 0) {
+        WB_NOTE("wc_xmss_hash n=32 arm failed");
+        wb_fail = 1;
+    }
+    state.params = &paramsPartial;
+    state.ret = 0;
+    wc_xmss_hash(&state, data, 16, hashOut);
+    if (state.ret != 0) {
+        WB_NOTE("wc_xmss_hash n=24 arm failed");
+        wb_fail = 1;
+    }
+    state.params = &paramsFull;
+
+    /* Lines 1033-1035 / 1218-1220: wc_xmss_rand_hash() /
+     * wc_xmss_rand_hash_lr()'s "params->n == XMSS_SHA256_32_N" operand. */
+    state.ret = 0;
+    XMEMSET(&addr, 0, sizeof(addr));
+    wc_xmss_rand_hash(&state, data, pk_seed, addr, hashOut);
+    if (state.ret != 0) {
+        WB_NOTE("wc_xmss_rand_hash n=32 arm failed");
+        wb_fail = 1;
+    }
+    state.params = &paramsPartial;
+    state.ret = 0;
+    XMEMSET(&addr, 0, sizeof(addr));
+    wc_xmss_rand_hash(&state, data, pk_seed, addr, hashOut);
+    if (state.ret != 0) {
+        WB_NOTE("wc_xmss_rand_hash n=24 arm failed");
+        wb_fail = 1;
+    }
+    state.params = &paramsFull;
+
+    state.ret = 0;
+    XMEMSET(&addr, 0, sizeof(addr));
+    wc_xmss_rand_hash_lr(&state, data, data + 32, pk_seed, addr, hashOut);
+    if (state.ret != 0) {
+        WB_NOTE("wc_xmss_rand_hash_lr n=32 arm failed");
+        wb_fail = 1;
+    }
+    state.params = &paramsPartial;
+    state.ret = 0;
+    XMEMSET(&addr, 0, sizeof(addr));
+    wc_xmss_rand_hash_lr(&state, data, data + 24, pk_seed, addr, hashOut);
+    if (state.ret != 0) {
+        WB_NOTE("wc_xmss_rand_hash_lr n=24 arm failed");
+        wb_fail = 1;
+    }
+    state.params = &paramsFull;
+
+#ifndef WOLFSSL_XMSS_VERIFY_ONLY
+    /* Lines 1813-1815: wc_xmss_wots_gen_pk(). */
+    state.ret = 0;
+    XMEMSET(&addr, 0, sizeof(addr));
+    XMEMSET(pkBuf, 0, sizeof(pkBuf));
+    wc_xmss_wots_gen_pk(&state, sk_seed, pk_seed, addr, pkBuf);
+    if (state.ret != 0) {
+        WB_NOTE("wc_xmss_wots_gen_pk n=32 arm failed");
+        wb_fail = 1;
+    }
+    state.params = &paramsPartial;
+    state.ret = 0;
+    XMEMSET(&addr, 0, sizeof(addr));
+    XMEMSET(pkBuf, 0, sizeof(pkBuf));
+    wc_xmss_wots_gen_pk(&state, sk_seed, pk_seed, addr, pkBuf);
+    if (state.ret != 0) {
+        WB_NOTE("wc_xmss_wots_gen_pk n=24 arm failed");
+        wb_fail = 1;
+    }
+    state.params = &paramsFull;
+
+    /* Lines 1884-1886: wc_xmss_wots_sign(). */
+    state.ret = 0;
+    XMEMSET(&addr, 0, sizeof(addr));
+    XMEMSET(sigBuf, 0, sizeof(sigBuf));
+    wc_xmss_wots_sign(&state, data, sk_seed, pk_seed, addr, sigBuf);
+    if (state.ret != 0) {
+        WB_NOTE("wc_xmss_wots_sign n=32 arm failed");
+        wb_fail = 1;
+    }
+    state.params = &paramsPartial;
+    state.ret = 0;
+    XMEMSET(&addr, 0, sizeof(addr));
+    XMEMSET(sigBuf, 0, sizeof(sigBuf));
+    wc_xmss_wots_sign(&state, data, sk_seed, pk_seed, addr, sigBuf);
+    if (state.ret != 0) {
+        WB_NOTE("wc_xmss_wots_sign n=24 arm failed");
+        wb_fail = 1;
+    }
+    state.params = &paramsFull;
+#else
+    WB_NOTE("WOLFSSL_XMSS_VERIFY_ONLY: wots_gen_pk/wots_sign (1813, 1884) "
+        "not compiled in; skipped");
+    XMEMSET(sigBuf, 0, sizeof(sigBuf));
+#endif /* !WOLFSSL_XMSS_VERIFY_ONLY */
+
+    /* Lines 1952-1954: wc_xmss_wots_pk_from_sig(). sigBuf holds whatever
+     * the wots_sign call above produced (or zeros, in VERIFY_ONLY builds);
+     * this decision's independence only needs the call to complete without
+     * a digest failure, not a cryptographically valid pk. */
+    state.ret = 0;
+    XMEMSET(&addr, 0, sizeof(addr));
+    XMEMSET(pkBuf, 0, sizeof(pkBuf));
+    wc_xmss_wots_pk_from_sig(&state, sigBuf, data, pk_seed, addr, pkBuf);
+    if (state.ret != 0) {
+        WB_NOTE("wc_xmss_wots_pk_from_sig n=32 arm failed");
+        wb_fail = 1;
+    }
+    state.params = &paramsPartial;
+    state.ret = 0;
+    XMEMSET(&addr, 0, sizeof(addr));
+    XMEMSET(pkBuf, 0, sizeof(pkBuf));
+    wc_xmss_wots_pk_from_sig(&state, sigBuf, data, pk_seed, addr, pkBuf);
+    if (state.ret != 0) {
+        WB_NOTE("wc_xmss_wots_pk_from_sig n=24 arm failed");
+        wb_fail = 1;
+    }
+    state.params = &paramsFull;
+
+    /* Lines 2022-2024 / 2053-2055: wc_xmss_ltree() - the same decision
+     * appears twice (once to prime the cached hash state, once inside the
+     * len-reduction loop); one call exercises both occurrences. pkBuf is
+     * used purely as WOTS+-shaped scratch input. */
+    state.ret = 0;
+    XMEMSET(&addr, 0, sizeof(addr));
+    wc_xmss_ltree(&state, pkBuf, pk_seed, addr, hashOut);
+    if (state.ret != 0) {
+        WB_NOTE("wc_xmss_ltree n=32 arm failed");
+        wb_fail = 1;
+    }
+    state.params = &paramsPartial;
+    state.ret = 0;
+    XMEMSET(&addr, 0, sizeof(addr));
+    wc_xmss_ltree(&state, pkBuf, pk_seed, addr, hashOut);
+    if (state.ret != 0) {
+        WB_NOTE("wc_xmss_ltree n=24 arm failed");
+        wb_fail = 1;
+    }
+    state.params = &paramsFull;
+
+    wb_state_free(&state);
+    WB_NOTE("hash-family n-operand independence pairs exercised");
+}
+
+/********************************************
+ * 1623, 1697: WOTS+ chain functions'
+ *   "for (i = start+1; i < (start+steps) && i < XMSS_WOTS_W; i++)"
+ * condIndex 1 ("i < XMSS_WOTS_W"). Calling with start=0, steps well beyond
+ * XMSS_WOTS_W (16) keeps "i < start+steps" true for the whole loop, so it
+ * is "i < XMSS_WOTS_W" alone that is true for i=1..15 and false at i=16 -
+ * both sides of that one operand, shown within this single call.
+ ********************************************/
+static void wb_wots_chain_loop(void)
+{
+    /* Line 1623: wc_xmss_chain_sha256_32() - fixed SHA-256/32-byte path. */
+    {
+        XmssParams params;
+        XmssState state;
+        ALIGN16 byte addrBuf[WC_XMSS_ADDR_LEN + 8];
+        byte data[32];
+        byte pkseed[32];
+        byte hashOut[32];
+        int i;
+
+        wb_params_init(&params, WC_HASH_TYPE_SHA256, 32, 32, 4, 1, 4, 0);
+        if (wb_state_init(&state, &params) != 0) {
+            WB_NOTE("wots chain loop: SHA-256 state init failed; skipped");
+        }
+        else {
+            for (i = 0; i < 32; i++) {
+                data[i] = (byte)i;
+                pkseed[i] = (byte)(0x50 + i);
+            }
+            XMEMSET(addrBuf, 0, sizeof(addrBuf));
+            state.ret = 0;
+            wc_xmss_chain_sha256_32(&state, data, 0, 40, pkseed, addrBuf,
+                hashOut);
+            if (state.ret != 0) {
+                WB_NOTE("wc_xmss_chain_sha256_32 start/steps run failed");
+                wb_fail = 1;
+            }
+            wb_state_free(&state);
+        }
+    }
+
+    /* Line 1697: wc_xmss_chain() - generic path. Using SHA-512 (n=64)
+     * guarantees this is NOT the sha256_32-specific fast path, so the
+     * generic wc_xmss_chain() implementation is what actually runs. */
+#ifdef WC_XMSS_SHA512
+    {
+        XmssParams params;
+        XmssState state;
+        ALIGN16 byte addrBuf[WC_XMSS_ADDR_LEN + 8];
+        byte data[64];
+        byte pkseed[64];
+        byte hashOut[64];
+        int i;
+
+        wb_params_init(&params, WC_HASH_TYPE_SHA512, 64, 64, 4, 1, 4, 0);
+        if (wb_state_init(&state, &params) != 0) {
+            WB_NOTE("wots chain loop: SHA-512 state init failed; skipped");
+        }
+        else {
+            for (i = 0; i < 64; i++) {
+                data[i] = (byte)i;
+                pkseed[i] = (byte)(0x60 + i);
+            }
+            XMEMSET(addrBuf, 0, sizeof(addrBuf));
+            state.ret = 0;
+            wc_xmss_chain(&state, data, 0, 40, pkseed, addrBuf, hashOut);
+            if (state.ret != 0) {
+                WB_NOTE("wc_xmss_chain start/steps run failed");
+                wb_fail = 1;
+            }
+            wb_state_free(&state);
+        }
+    }
+#else
+    WB_NOTE("WC_XMSS_SHA512 not compiled in; generic wc_xmss_chain (line "
+        "1697) arm skipped");
+#endif
+}
+
+#ifndef WOLFSSL_XMSS_VERIFY_ONLY
+/********************************************
+ * 2846: wc_xmss_bds_next_idx()'s "if ((hsk > 0) && (i == 3))".
+ * hsk = sub_h - bds_k. Direct calls with offset=0 (so the function's
+ * internal "while (o >= 1...)" loop never runs) keep this test isolated to
+ * just the targeted if.
+ ********************************************/
+static void wb_bds_next_idx(void)
+{
+    XmssParams paramsPos;   /* bds_k=0  -> hsk=sub_h=4 > 0 */
+    XmssParams paramsZero;  /* bds_k=4  -> hsk=0 */
+    XmssState state;
+    BdsState bdsPos[1];
+    BdsState bdsZero[1];
+    byte skPos[2048];
+    byte skZero[2048];
+    byte sk_seed[32];
+    byte pk_seed[32];
+    HashAddress addr;
+    word8 height[8];
+    word8 offset;
+    byte* sp;
+
+    wb_params_init(&paramsPos, WC_HASH_TYPE_SHA256, 32, 32, 4, 1, 4, 0);
+    wb_params_init(&paramsZero, WC_HASH_TYPE_SHA256, 32, 32, 4, 1, 4, 4);
+
+    if (wb_state_init(&state, &paramsPos) != 0) {
+        WB_NOTE("bds_next_idx: state init failed; skipped");
+        return;
+    }
+    XMEMSET(sk_seed, 0x33, sizeof(sk_seed));
+    XMEMSET(pk_seed, 0x44, sizeof(pk_seed));
+    XMEMSET(skPos, 0, sizeof(skPos));
+    XMEMSET(skZero, 0, sizeof(skZero));
+
+    /* hsk > 0, i == 3: both operands true -> the guarded copy runs. */
+    if (wc_xmss_bds_state_load(&state, skPos, bdsPos, NULL) == 0) {
+        XMEMSET(&addr, 0, sizeof(addr));
+        XMEMSET(height, 0, sizeof(height));
+        offset = 0;
+        sp = state.stack;
+        state.ret = 0;
+        wc_xmss_bds_next_idx(&state, &bdsPos[0], sk_seed, pk_seed, addr, 3,
+            height, &offset, &sp);
+        if (state.ret != 0) {
+            WB_NOTE("bds_next_idx hsk>0,i==3 call failed");
+            wb_fail = 1;
+        }
+    }
+    else {
+        WB_NOTE("bds_next_idx: bds_state_load (hsk>0) failed; skipped");
+    }
+
+    /* hsk > 0, i != 3: condIndex1 false, condIndex0 held true. */
+    if (wc_xmss_bds_state_load(&state, skPos, bdsPos, NULL) == 0) {
+        XMEMSET(&addr, 0, sizeof(addr));
+        XMEMSET(height, 0, sizeof(height));
+        offset = 0;
+        sp = state.stack;
+        state.ret = 0;
+        wc_xmss_bds_next_idx(&state, &bdsPos[0], sk_seed, pk_seed, addr, 5,
+            height, &offset, &sp);
+        if (state.ret != 0) {
+            WB_NOTE("bds_next_idx hsk>0,i!=3 call failed");
+            wb_fail = 1;
+        }
+    }
+
+    /* hsk == 0: condIndex0 false (masks the AND) with i == 3 held true. */
+    state.params = &paramsZero;
+    if (wc_xmss_bds_state_load(&state, skZero, bdsZero, NULL) == 0) {
+        XMEMSET(&addr, 0, sizeof(addr));
+        XMEMSET(height, 0, sizeof(height));
+        offset = 0;
+        sp = state.stack;
+        state.ret = 0;
+        wc_xmss_bds_next_idx(&state, &bdsZero[0], sk_seed, pk_seed, addr, 3,
+            height, &offset, &sp);
+        if (state.ret != 0) {
+            WB_NOTE("bds_next_idx hsk==0,i==3 call failed");
+            wb_fail = 1;
+        }
+    }
+    else {
+        WB_NOTE("bds_next_idx: bds_state_load (hsk==0) failed; skipped");
+    }
+    state.params = &paramsPos;
+
+    wb_state_free(&state);
+}
+
+/********************************************
+ * 3212: wc_xmss_bds_auth_path()'s
+ *   "if ((bds->keep == NULL) || (bds->authPath == NULL))"
+ * 3251: same function's "if ((tau < hs - 1) && (parent == 0))", where tau
+ * and parent come from wc_xmss_lowest_zero_bit_index(leafIdx, hs, &parent).
+ * For hs=4: leafIdx=1 -> tau=1,parent=0 (both true); leafIdx=5 -> tau=1,
+ * parent=1 (condIndex1 false, condIndex0 held true); leafIdx=7 -> tau=3,
+ * parent=0 (condIndex0 false, masking condIndex1). bds_k=0 keeps hsk=4
+ * (>= all tau values used here), so the "i < hsk" arm of the trailing
+ * per-height loop is always taken and bds->retain is never dereferenced.
+ ********************************************/
+static void wb_bds_auth_path(void)
+{
+    XmssParams params;
+    XmssState state;
+    BdsState bds[1];
+    byte skBuf[2048];
+    byte sk_seed[32];
+    byte pk_seed[32];
+    HashAddress addr;
+
+    wb_params_init(&params, WC_HASH_TYPE_SHA256, 32, 32, 4, 1, 4, 0);
+
+    if (wb_state_init(&state, &params) != 0) {
+        WB_NOTE("bds_auth_path: state init failed; skipped");
+        return;
+    }
+    XMEMSET(sk_seed, 0x55, sizeof(sk_seed));
+    XMEMSET(pk_seed, 0x66, sizeof(pk_seed));
+
+    /* Line 3212, "keep == NULL" true side (authPath left valid). */
+    XMEMSET(skBuf, 0, sizeof(skBuf));
+    if (wc_xmss_bds_state_load(&state, skBuf, bds, NULL) == 0) {
+        byte* savedKeep = bds[0].keep;
+
+        bds[0].keep = NULL;
+        state.ret = 0;
+        XMEMSET(&addr, 0, sizeof(addr));
+        wc_xmss_bds_auth_path(&state, &bds[0], 1, sk_seed, pk_seed, addr);
+        if (state.ret != WC_NO_ERR_TRACE(WC_FAILURE)) {
+            WB_NOTE("bds_auth_path keep==NULL did not report WC_FAILURE");
+            wb_fail = 1;
+        }
+        bds[0].keep = savedKeep;
+    }
+    else {
+        WB_NOTE("bds_auth_path: bds_state_load failed; skipped");
+    }
+
+    /* Line 3212, "authPath == NULL" true side (keep left valid). */
+    XMEMSET(skBuf, 0, sizeof(skBuf));
+    if (wc_xmss_bds_state_load(&state, skBuf, bds, NULL) == 0) {
+        byte* savedAuth = bds[0].authPath;
+
+        bds[0].authPath = NULL;
+        state.ret = 0;
+        XMEMSET(&addr, 0, sizeof(addr));
+        wc_xmss_bds_auth_path(&state, &bds[0], 1, sk_seed, pk_seed, addr);
+        if (state.ret != WC_NO_ERR_TRACE(WC_FAILURE)) {
+            WB_NOTE("bds_auth_path authPath==NULL did not report "
+                "WC_FAILURE");
+            wb_fail = 1;
+        }
+        bds[0].authPath = savedAuth;
+    }
+
+    /* Baseline + line 3251: leafIdx=1 -> tau=1,parent=0 -> both operands
+     * true (also exercises line 3212's false side with both pointers
+     * valid). */
+    XMEMSET(skBuf, 0, sizeof(skBuf));
+    if (wc_xmss_bds_state_load(&state, skBuf, bds, NULL) == 0) {
+        state.ret = 0;
+        XMEMSET(&addr, 0, sizeof(addr));
+        wc_xmss_bds_auth_path(&state, &bds[0], 1, sk_seed, pk_seed, addr);
+        if (state.ret != 0) {
+            WB_NOTE("bds_auth_path leafIdx=1 call failed");
+            wb_fail = 1;
+        }
+    }
+
+    /* Line 3251: leafIdx=5 -> tau=1,parent=1 -> condIndex1 false,
+     * condIndex0 held true. */
+    XMEMSET(skBuf, 0, sizeof(skBuf));
+    if (wc_xmss_bds_state_load(&state, skBuf, bds, NULL) == 0) {
+        state.ret = 0;
+        XMEMSET(&addr, 0, sizeof(addr));
+        wc_xmss_bds_auth_path(&state, &bds[0], 5, sk_seed, pk_seed, addr);
+        if (state.ret != 0) {
+            WB_NOTE("bds_auth_path leafIdx=5 call failed");
+            wb_fail = 1;
+        }
+    }
+
+    /* Line 3251: leafIdx=7 -> tau=3,parent=0 -> condIndex0 false (masks
+     * condIndex1). */
+    XMEMSET(skBuf, 0, sizeof(skBuf));
+    if (wc_xmss_bds_state_load(&state, skBuf, bds, NULL) == 0) {
+        state.ret = 0;
+        XMEMSET(&addr, 0, sizeof(addr));
+        wc_xmss_bds_auth_path(&state, &bds[0], 7, sk_seed, pk_seed, addr);
+        if (state.ret != 0) {
+            WB_NOTE("bds_auth_path leafIdx=7 call failed");
+            wb_fail = 1;
+        }
+    }
+
+    wb_state_free(&state);
+}
+
+/********************************************
+ * 3651: wc_xmssmt_keygen()'s (non-SMALL, active variant)
+ *   "for (i = 0; (ret == 0) && (i < params->d - 1); i++)"
+ * A real d=2 keygen naturally exercises "i < d - 1" true (i=0) then false
+ * (i=1) with ret==0 throughout a successful call.
+ *
+ * 3956, 3985-3987: wc_xmssmt_sign_next_idx() (static, only reachable
+ * through wc_xmssmt_sign()) - repeatedly signing every valid index of a
+ * small d=2 tree drives its internal per-subtree bookkeeping (including the
+ * subtree-boundary crossings every 2^sub_h signs) through many operand
+ * combinations, with ret==0 throughout a successful run.
+ *
+ * 4379: wc_xmssmt_verify()'s "for (i = 1; (ret==0) && (i < params->d);
+ * i++)" - a d=2 verify naturally exercises "i < d" true (i=1) then false
+ * (i=2).
+ ********************************************/
+static void wb_full_cycle_d2(void)
+{
+    XmssParams params;
+    XmssState state;
+    byte seed[3 * 32];
+    byte sk[8192];
+    byte pk[160];
+    byte sig[8192];
+    static const byte msg[] = "xmss whitebox d2 message";
+    int ret;
+    int i;
+
+    wb_params_init(&params, WC_HASH_TYPE_SHA256, 32, 32, 4, 2, 4, 0);
+
+    if (wb_state_init(&state, &params) != 0) {
+        WB_NOTE("d2 cycle: state init failed; skipped");
+        return;
+    }
+
+    XMEMSET(seed, 0x22, sizeof(seed));
+    XMEMSET(sk, 0, sizeof(sk));
+    XMEMSET(pk, 0, sizeof(pk));
+
+    ret = wc_xmssmt_keygen(&state, seed, sk, pk);
+    if (ret != 0) {
+        WB_NOTE("d2 cycle: keygen failed; skipped");
+        wb_state_free(&state);
+        return;
+    }
+
+    /* Sign through every valid index (0..2^h-2 == 14): exercises
+     * wc_xmssmt_sign_next_idx()'s internals across both subtrees. */
+    for (i = 0; i < 15; i++) {
+        XMEMSET(sig, 0, sizeof(sig));
+        ret = wc_xmssmt_sign(&state, msg, (word32)sizeof(msg), sk, sig);
+        if (ret != 0) {
+            WB_NOTE("d2 cycle: sign failed before exhaustion");
+            wb_fail = 1;
+            break;
+        }
+
+        if (i == 14) {
+            /* Last valid index: verify exercises the full d=2 loop at
+             * line 4379 (i=1<2 true, then i=2<2 false), ret==0
+             * throughout. */
+            ret = wc_xmssmt_verify(&state, msg, (word32)sizeof(msg), sig,
+                pk);
+            if (ret != 0) {
+                WB_NOTE("d2 cycle: verify of a good signature failed");
+                wb_fail = 1;
+            }
+        }
+    }
+
+    wb_state_free(&state);
+}
+
+/********************************************
+ * 4070: wc_xmssmt_sign()'s "if ((ret == 0) && xmss_idx_invalid(idx, h))".
+ * 4087: same function's
+ *   "if ((ret == 0) && (idx < (((XmssIdx)1 << h) - 1)))" - given the
+ *   upstream invalid-index check at 4070 already rejects any idx that
+ *   would make this false (the largest idx that can reach here is
+ *   2^h - 2, and 2^h - 2 < 2^h - 1 always), condIndex1's false side is
+ *   provably unreachable while ret==0; only condIndex0 is closed here
+ *   (see SKIP note in the final report).
+ * 4121: wc_xmss_sigsleft()'s
+ *   "if ((ret == 0) && (WC_IDX_INVALID(idx, params->idx_len, params->h)))".
+ * 4391: wc_xmssmt_verify()'s "if ((ret == 0) && (XMEMCMP(node, pub_root, n)
+ *   != 0))" - forced both ways directly (valid vs. corrupted public key).
+ *
+ * A small d=1 tree is signed through every valid index until natural
+ * exhaustion (idx reaches 2^h - 1), which is exactly when line 4070's
+ * "xmss_idx_invalid" operand flips true (having been false on every prior,
+ * successful sign) - masking line 4087's "ret == 0" operand to false on
+ * that same call. A separate, cheap direct call to wc_xmss_sigsleft() with
+ * an idx_len wc_xmss's dual-width WC_IDX_DECODE doesn't recognize (2, vs.
+ * the valid 3/4/5/8) forces "ret == 0" false at line 4121 without ever
+ * evaluating WC_IDX_INVALID - the complementary independence pair to the
+ * exhausted-key call (which shows WC_IDX_INVALID's true side with ret==0
+ * true) and the fresh-key call (WC_IDX_INVALID's false side, ret==0 true).
+ ********************************************/
+static void wb_full_cycle_d1(void)
+{
+    XmssParams params;
+    XmssState state;
+    byte seed[3 * 32];
+    byte sk[2048];
+    byte pk[160];
+    byte sig[4096];
+    static const byte msg[] = "xmss whitebox d1 message";
+    int ret;
+    int i;
+    int exhausted = 0;
+
+    wb_params_init(&params, WC_HASH_TYPE_SHA256, 32, 32, 4, 1, 4, 0);
+
+    if (wb_state_init(&state, &params) != 0) {
+        WB_NOTE("d1 cycle: state init failed; skipped");
+        return;
+    }
+
+    XMEMSET(seed, 0x11, sizeof(seed));
+    XMEMSET(sk, 0, sizeof(sk));
+    XMEMSET(pk, 0, sizeof(pk));
+
+    ret = wc_xmssmt_keygen(&state, seed, sk, pk);
+    if (ret != 0) {
+        WB_NOTE("d1 cycle: keygen failed; skipped");
+        wb_state_free(&state);
+        return;
+    }
+
+    /* wc_xmss_sigsleft() on a fresh, unexhausted key: line 4121's
+     * WC_IDX_INVALID false side, ret == 0 true. */
+    ret = wc_xmss_sigsleft(&params, sk);
+    if (ret != 1) {
+        WB_NOTE("d1 cycle: sigsleft on a fresh key did not report sigs "
+            "left");
+        wb_fail = 1;
+    }
+
+    /* Sign through every valid index (0..2^h-2 == 14). */
+    for (i = 0; i < 15; i++) {
+        XMEMSET(sig, 0, sizeof(sig));
+        ret = wc_xmssmt_sign(&state, msg, (word32)sizeof(msg), sk, sig);
+        if (ret != 0) {
+            WB_NOTE("d1 cycle: sign failed before exhaustion");
+            wb_fail = 1;
+            break;
+        }
+
+        if (i == 0) {
+            /* Genuine verify: line 4391's XMEMCMP==0 (false) side, and
+             * the successful d=1 loop path at line 4379. */
+            ret = wc_xmssmt_verify(&state, msg, (word32)sizeof(msg), sig,
+                pk);
+            if (ret != 0) {
+                WB_NOTE("d1 cycle: verify of a good signature failed");
+                wb_fail = 1;
+            }
+
+            /* Corrupt the public root and re-verify: line 4391's
+             * XMEMCMP!=0 (true) side, ret==0 up to that point. */
+            {
+                byte badPk[160];
+
+                XMEMCPY(badPk, pk, sizeof(pk));
+                badPk[0] ^= 0xFFU;
+                ret = wc_xmssmt_verify(&state, msg, (word32)sizeof(msg),
+                    sig, badPk);
+                if (ret != WC_NO_ERR_TRACE(SIG_VERIFY_E)) {
+                    WB_NOTE("d1 cycle: corrupted-pk verify did not fail "
+                        "as expected");
+                    wb_fail = 1;
+                }
+            }
+        }
+    }
+
+    /* sk's index is now 2^h - 1 == 15. This sign attempt hits line 4070's
+     * "xmss_idx_invalid(idx,h)" true side (ret==0 up to that point),
+     * forcing ret = KEY_EXHAUSTED_E, which (masking) drives line 4087's
+     * "ret == 0" operand to false. */
+    ret = wc_xmssmt_sign(&state, msg, (word32)sizeof(msg), sk, sig);
+    if (ret == WC_NO_ERR_TRACE(KEY_EXHAUSTED_E)) {
+        exhausted = 1;
+    }
+    else {
+        WB_NOTE("d1 cycle: key was not reported exhausted as expected");
+        wb_fail = 1;
+    }
+
+    /* wc_xmss_sigsleft(): line 4121's WC_IDX_INVALID true side, ret == 0
+     * true. Craft an sk whose encoded idx is exactly 2^h - 1 == 15 (the
+     * smallest value for which (idx+1)>>h != 0) directly, rather than
+     * reusing the just-exhausted sk above: wc_xmssmt_sign()'s exhaustion
+     * handling XMEMSETs the index field to all-0xFF, which as an encoded
+     * 32-bit value (0xFFFFFFFF) wraps back to looking "valid" under
+     * IDX32_INVALID's "(idx+1)>>h" arithmetic (idx+1 overflows to 0) - a
+     * real quirk of that cleanup path, but not what this test is after. */
+    if (exhausted) {
+        byte idxSk[2048];
+
+        XMEMCPY(idxSk, sk, sizeof(idxSk));
+        idxSk[0] = 0x00;
+        idxSk[1] = 0x00;
+        idxSk[2] = 0x00;
+        idxSk[3] = 0x0F; /* idx = 15 = 2^h - 1, h = 4 */
+        ret = wc_xmss_sigsleft(&params, idxSk);
+        if (ret != 0) {
+            WB_NOTE("d1 cycle: sigsleft with idx==2^h-1 unexpectedly "
+                "reported sigs left");
+            wb_fail = 1;
+        }
+    }
+
+    wb_state_free(&state);
+
+    /* Line 4121, "ret == 0" false side: idx_len=2 matches neither
+     * WC_IDX_DECODE's 32-bit arm (3 or 4 bytes) nor its 64-bit arm (5 or
+     * 8 bytes), so decode sets ret = NOT_COMPILED_IN and WC_IDX_INVALID is
+     * never evaluated - independent of any sk content. */
+    {
+        XmssParams badLenParams;
+        byte dummySk[8];
+
+        wb_params_init(&badLenParams, WC_HASH_TYPE_SHA256, 32, 32, 4, 1, 2,
+            0);
+        XMEMSET(dummySk, 0, sizeof(dummySk));
+        ret = wc_xmss_sigsleft(&badLenParams, dummySk);
+        if (ret != 0) {
+            WB_NOTE("d1 cycle: sigsleft with an unsupported idx_len "
+                "unexpectedly reported sigs left");
+            wb_fail = 1;
+        }
+    }
+}
+#else /* WOLFSSL_XMSS_VERIFY_ONLY */
+static void wb_bds_next_idx(void)
+{
+    WB_NOTE("WOLFSSL_XMSS_VERIFY_ONLY: signing-side BDS helpers not "
+        "compiled in; wb_bds_next_idx skipped");
+}
+static void wb_bds_auth_path(void)
+{
+    WB_NOTE("WOLFSSL_XMSS_VERIFY_ONLY: signing-side BDS helpers not "
+        "compiled in; wb_bds_auth_path skipped");
+}
+static void wb_full_cycle_d2(void)
+{
+    WB_NOTE("WOLFSSL_XMSS_VERIFY_ONLY: keygen/sign not compiled in; "
+        "wb_full_cycle_d2 skipped");
+}
+static void wb_full_cycle_d1(void)
+{
+    WB_NOTE("WOLFSSL_XMSS_VERIFY_ONLY: keygen/sign not compiled in; "
+        "wb_full_cycle_d1 skipped");
+}
+#endif /* !WOLFSSL_XMSS_VERIFY_ONLY */
+
+#else /* WOLFSSL_HAVE_XMSS */
+
+static void wb_hash_family_pairs(void)
+{
+    WB_NOTE("WOLFSSL_HAVE_XMSS not compiled in; skipped");
+}
+static void wb_wots_chain_loop(void)
+{
+    WB_NOTE("WOLFSSL_HAVE_XMSS not compiled in; skipped");
+}
+static void wb_bds_next_idx(void)
+{
+    WB_NOTE("WOLFSSL_HAVE_XMSS not compiled in; skipped");
+}
+static void wb_bds_auth_path(void)
+{
+    WB_NOTE("WOLFSSL_HAVE_XMSS not compiled in; skipped");
+}
+static void wb_full_cycle_d2(void)
+{
+    WB_NOTE("WOLFSSL_HAVE_XMSS not compiled in; skipped");
+}
+static void wb_full_cycle_d1(void)
+{
+    WB_NOTE("WOLFSSL_HAVE_XMSS not compiled in; skipped");
+}
+
+#endif /* WOLFSSL_HAVE_XMSS */
+
 int main(void)
 {
     printf("wc_xmss_impl.c white-box supplement\n");
     wb_run();
+    wb_hash_family_pairs();
+    wb_wots_chain_loop();
+    wb_bds_next_idx();
+    wb_bds_auth_path();
+    wb_full_cycle_d2();
+    wb_full_cycle_d1();
     printf("done (%s)\n", wb_fail ? "with failures" : "ok");
     return 0;
 }
